rename files and directory
https://coveord.atlassian.net/browse/KIT-1131
