+++ packages/headless/src/utils/query-expression/common/negatable.ts
