+++ packages/headless/src/utils/query-expression/common/part.ts
