+++ packages/headless/src/utils/query-expression/query-expression.ts
@@ -54,102 +54,100 @@ export {
 /**
  * A utility to help build query expressions.
  */
-export interface ExpressionBuilder {
+export interface QueryExpression {
   /**
    * Adds an expression containing terms to match. Terms can be in any order, and may also be expanded with stemming.
    *
    * @param expression - A keyword expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addKeyword(expression: KeywordExpression): ExpressionBuilder;
+  addKeyword(expression: KeywordExpression): QueryExpression;
 
   /**
    * Adds an expression that returns all of the items in which the specified `startTerm` appears no more than `maxKeywordsBetween` from the endTerm, for each element in `otherTerms`.
    *
    * @param expression - A near expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addNear(expression: NearExpression): ExpressionBuilder;
+  addNear(expression: NearExpression): QueryExpression;
 
   /**
    * Adds an expression that must appear in its entirety at least once for an item to be returned.
    *
    * @param expression - An exact match expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addExactMatch(expression: ExactMatchExpression): ExpressionBuilder;
+  addExactMatch(expression: ExactMatchExpression): QueryExpression;
 
   /**
    * Adds an expression returning all items where the defined field exists.
    *
    * @param expression - A field exists expressions.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addFieldExists(expression: FieldExistsExpression): ExpressionBuilder;
+  addFieldExists(expression: FieldExistsExpression): QueryExpression;
 
   /**
    * Adds an expression that uses an `operator` to compare a string `field` against certain `values`.
    * Returns all of the items for which the expression evaluates to true.
    *
    * @param expression - A string field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addStringField(expression: StringFieldExpression): ExpressionBuilder;
+  addStringField(expression: StringFieldExpression): QueryExpression;
 
   /**
    * Adds an expression that uses an `operator` to compare a string facet `field` to a `value`.
    * Returns all of the items for which the expression evaluates to true.
    *
    * @param expression - A string facet field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addStringFacetField(
-    expression: StringFacetFieldExpression
-  ): ExpressionBuilder;
+  addStringFacetField(expression: StringFacetFieldExpression): QueryExpression;
 
   /**
    * Adds an expression that uses an `operator` to compare a numeric `field` to a `value`.
    * Returns all of the items for which the expression evaluates to true.
    *
    * @param expression - A numeric field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addNumericField(expression: NumericFieldExpression): ExpressionBuilder;
+  addNumericField(expression: NumericFieldExpression): QueryExpression;
 
   /**
    * Adds an expression that returns all items for which the `value` of the numeric `field` is within the defined range.
    *
    * @param expression - A numeric field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
   addNumericRangeField(
     expression: NumericRangeFieldExpression
-  ): ExpressionBuilder;
+  ): QueryExpression;
 
   /**
    * Adds an expression that uses an `operator` to compare a date `field` to a `value`.
    * Returns all of the items for which the expression evaluates to true.
    *
    * @param expression - A date field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addDateField(expression: DateFieldExpression): ExpressionBuilder;
+  addDateField(expression: DateFieldExpression): QueryExpression;
 
   /**
    * Adds an expression that returns all items for which the `value` of the date `field` is within the defined range.
    *
    * @param expression - A numeric field expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addDateRangeField(expression: DateRangeFieldExpression): ExpressionBuilder;
+  addDateRangeField(expression: DateRangeFieldExpression): QueryExpression;
 
   /**
    * Adds an expression that invokes a query extension.
    *
    * @param expression - A query extension expression.
-   * @returns The `ExpressionBuilder` instance.
+   * @returns The `QueryExpression` instance.
    */
-  addQueryExtension(expression: QueryExtensionExpression): ExpressionBuilder;
+  addQueryExtension(expression: QueryExtensionExpression): QueryExpression;
 
   /**
    * Joins all expressions using the configured boolean operator.
@@ -164,7 +162,7 @@ type BooleanOperator = 'and' | 'or';
 /**
  * The expression builder options.
  */
-export interface ExpressionBuilderOptions {
+export interface QueryExpressionOptions {
   /**
    * The boolean operator to join individual expressions with.
    */
@@ -172,14 +170,14 @@ export interface ExpressionBuilderOptions {
 }
 
 /**
- * Creates an `ExpressionBuilder` instance.
+ * Creates an `QueryExpression` instance.
  *
  * @param config - The expression builder options.
- * @returns An `ExpressionBuilder` instance.
+ * @returns An `QueryExpression` instance.
  */
 export function buildQueryExpression(
-  config: ExpressionBuilderOptions
-): ExpressionBuilder {
+  config: QueryExpressionOptions
+): QueryExpression {
   const parts: Part[] = [];
 
   return {
