+++ packages/headless/src/utils/query-expression/string-facet-field/string-facet-field.ts
